feat(core): Extract `sampleTransaction` method out (#9136)
This extracts the `sampleTransaction` method out and exports this from
`core` so we can use it for POTEL (where we build transactions
differently).
